% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGroupedPValues.R
\name{combineGroupedPValues}
\alias{combineGroupedPValues}
\title{Combine grouped p-values}
\usage{
combineGroupedPValues(
  p.values,
  grouping,
  method = c("simes", "holm-min", "berger", "fisher", "pearson", "wilkinson",
    "stouffer"),
  weights = NULL,
  log.p = FALSE,
  min.n = 1,
  min.prop = 0.5,
  is.rle = FALSE
)
}
\arguments{
\item{p.values}{A numeric vector containing p-values for individual tests.}

\item{grouping}{A vector or factor of length equal to \code{p.values}, specifying the group to which each test is assigned.

Alternatively, if \code{is.rle=TRUE}, an \link{rle} object specifying consecutive entries of \code{effects} that are in the same group.}

\item{weights}{A numeric vector of length equal to \code{p.values}, containing a positive weight for each test.
Alternatively \code{NULL}, in which case equal weights are assigned to all tests.}

\item{log.p}{Logical scalar indicating whether the p-values in \code{p.values} are log-transformed.}

\item{is.rle}{Logical scalar indicating whether \code{grouping} is an \link{rle} object.}
}
\value{
A numeric vector containing the combined p-values, named according to the grouping level.
}
\description{
Combine p-values from grouped hypothesis tests using a variety of meta-analysis methods.
Each group of p-values is defined as those assigned to the same level of the grouping factor.
}
\examples{
p <- runif(10000)
g <- sample(100, 10000, replace=TRUE)

fish <- combineGroupedPValues(p, g, method="fisher")
hist(fish$p.value)

z <- combineGroupedPValues(p, g, method="stouffer", weights=rexp(10000))
hist(z$p.value)

simes <- combineGroupedPValues(p, g, method="simes")
hist(simes$p.value)

berger <- combineGroupedPValues(p, g, method="berger")
hist(berger$p.value)

}
\author{
Aaron Lun
}
