% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapod-package.R
\docType{package}
\name{metapod-package}
\alias{metapod}
\alias{metapod-package}
\title{metapod: Meta-Analyses on P-Values of Differential Analyses}
\description{
Implements a variety of methods for combining p-values in differential analyses of genome-scale datasets. Functions can combine p-values across different tests in the same analysis (e.g., genomic windows in ChIP-seq, exons in RNA-seq) or for corresponding tests across separate analyses (e.g., replicated comparisons, effect of different treatment conditions). Support is provided for handling log-transformed input p-values, missing values and weighting where appropriate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LTLA/metapod}
  \item Report bugs at \url{https://github.com/LTLA/metapod/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

}
