% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeGroupedDirection.R
\name{summarizeGroupedDirection}
\alias{summarizeGroupedDirection}
\title{Summarize overall direction of grouped tests}
\usage{
summarizeGroupedDirection(
  effects,
  grouping,
  influential,
  threshold = 0,
  is.rle = FALSE
)
}
\arguments{
\item{effects}{A numeric vector containing the effect size for each test.}

\item{grouping}{A vector of factor of length equal to \code{effects}, specifying the assigned group for each tests.

Alternatively, if \code{is.rle=TRUE}, an \link{rle} object specifying consecutive entries of \code{effects} that are in the same group.}

\item{influential}{A logical vector of length equal to \code{effects},
indicating whether each test is influential in its assigned group.}

\item{threshold}{Numeric scalar defining the threshold at which an effect is \dQuote{"up"} or \dQuote{"down"}.}

\item{is.rle}{Logical scalar indicating whether \code{grouping} is an \link{rle} object.}
}
\value{
A character vector of length equal to the number of groups.
Each entry can be:
\itemize{
\item \code{"up"}, if all influential tests have effects above \code{threshold}.
\item \code{"down"}, if all influential tests have effects below \code{threshold}.
\item \code{"none"}, if all influential tests have effects equal to \code{threshold}.
\item \code{"mixed"}, if there are influential tests with effects above and below \code{threshold}.
}
}
\description{
Summarize the overall direction of grouped tests in a meta-analysis,
based on the influential tests defined by one of the \code{grouped*} functions.
}
\details{
By focusing on only the influential tests, we avoid introducing unnecessary noise in the high-level direction summary.
For example, if we did our meta-analysis using \code{\link{parallelSimes}}, 
we are not particularly concerned about the direction of tests with large p-values;
so we just ignore them when summarizing the direction in this function.
Of course, the interpretation of \dQuote{influential} really depends on the choice of meta-analysis strategy.
}
\examples{
p <- rbeta(100, 0.5, 1)
eff <- rnorm(100)
g <- sample(20, 100, replace=TRUE)

out <- groupedSimes(p, g)
(dir <- summarizeGroupedDirection(eff, g, out$influential))
}
\seealso{
\code{\link{groupedSimes}} and related \code{grouped*} functions, to obtain \code{influential}.

\code{\link{summarizeParallelDirection}}, for the equivalent function based on parallel tests.
}
\author{
Aaron Lun
}
